Make build process more user-friendly:
1) The autogenerated files
  m4/libtool.m4
  m4/ltoptions.m4
  m4/ltversion.m4
  m4/lt~obsolete.m4
are removed from repository.

2) The m4/pandora_have_libmysqlclient.m4 script
is removed from repository because it supposes to be
in pandora-build project.

3) The mysql client library inclusion is removed from
query_log plugin source code because one shouldn't
depend on database layer.

4) The mysql client library inclusion is added in
mysql_client plugin because it doesn't build without that
header.

5) The mysql client library header files path is added
to CPPFLAGS if mysql_client plugin because it didn't
build without additional dancing.