Group: directive to the .spec file, boost 1.41 requirement
Old versions of rpmbuild(1) require a Group: directive to be included
in the main definition of the file. Add an Applications/Database group
directive.

rhel5 provides boost 1.41, required by Percona Playback, in the
boost141-* packages, instead of boost-*. Add a switch to select the
right requirements for building and running

rhel5 also needs a BuildRoot declaration in order to be able to use the
RPM_BUILD_ROOT internal variable later.