General log support.
General log files don't contain as much information as the slow log file
so use them only as a last option. You will only have the thread id and
the query.

Use this option --ignore-row-result-diffs to discard the number of
returned rows as errors, otherwise you are going to have a huge list of
problems.

Usage : --input-plugin general-log --general-log-file or
--general-log-stdin